header cleanup using IWYU tool (#1467)
* Forward declaration of structures in headers where possible

* Use M_snprinf instead of sprintf

* Remove D_INTMAX D_INTMIN

* Don't change libtextscreen files

* Add miniz and win32 include directories

+++ src/dstrings.c
@@ -67,7 +67,7 @@ char **endmsg[]=
 };
 
 // killough 1/18/98: remove hardcoded limit and replace with var (silly hack):
-const size_t NUM_QUITMESSAGES = sizeof(endmsg)/sizeof(*endmsg) - 1;
+const int NUM_QUITMESSAGES = sizeof(endmsg)/sizeof(*endmsg) - 1;
 
   
 //----------------------------------------------------------------------------
